Divide test_panels tests
